Class4_6/requirements.txt
adafruit-io==1.1.0
paho-mqtt==1.3.1
